requirements.txt
discord.py
llama-cpp-python
fastapi
pydantic
uvicorn
requests
